fix: clear the mechanical clippy backlog the lint gate rejects

+++ src/api/client/dispatch.rs
@@ -121,10 +121,10 @@ impl EClient {
             let parent_id = self.core.tracked_parent_id(update.order_id)
                 .unwrap_or(update.parent_id);
             wrapper.order_status(
-                update.order_id as i64, status, update.filled_qty as f64,
-                update.remaining_qty as f64, 0.0, update.perm_id, parent_id, 0.0, 0, "", 0.0,
+                update.order_id as i64, status, update.filled_qty,
+                update.remaining_qty, 0.0, update.perm_id, parent_id, 0.0, 0, "", 0.0,
             );
-            self.core.update_order_status(&self.shared, update.order_id, update.status, update.filled_qty as f64, update.remaining_qty as f64);
+            self.core.update_order_status(&self.shared, update.order_id, update.status, update.filled_qty, update.remaining_qty);
         }
 
         // Cancel rejects → error
